Extract EOLLogProcessor module
