+++ lib/line-builders.js
@@ -5,12 +5,6 @@ const {
   prettifyObject
 } = require('./utils')
 
-function appendEOL (lineParts, { EOL }) {
-  if (lineParts.length > 0) {
-    lineParts.push(EOL)
-  }
-}
-
 function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys, errorProps, messageKey }) {
   const errorProperties = errorProps.split(',')
 
@@ -37,7 +31,6 @@ function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys,
 }
 
 const builtInLineBuilders = {
-  appendEOL,
   appendObjectOrError
 }
