+++ lib/log-processors.js
@@ -9,9 +9,9 @@ const MessageLogProcessor = require('./processors/MessageLogProcessor')
 const MetadataLogProcessor = require('./processors/MetadataLogProcessor')
 const SemicolonLogProcessor = require('./processors/SemicolonLogProcessor')
 const TimeLogProcessor = require('./processors/TimeLogProcessor')
+const EOLLogProcessor = require('./processors/EOLLogProcessor')
 
 const {
-  appendEOL,
   appendObjectOrError
 } = builtInLineBuilders
 
@@ -27,12 +27,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class EOLLogProcessor extends LogProcessor {
-  build (lineParts, context) {
-    return appendEOL(lineParts, context)
   }
 }
