+++ lib/processors/EOLLogProcessor.js
@@ -1,0 +1,15 @@
+const LogProcessor = require('../LogProcessor')
+
+function appendEOL (lineParts, { EOL }) {
+  if (lineParts.length > 0) {
+    lineParts.push(EOL)
+  }
+}
+
+class EOLLogProcessor extends LogProcessor {
+  build (lineParts, context) {
+    return appendEOL(lineParts, context)
+  }
+}
+
+module.exports = EOLLogProcessor
